[Modify] move `PieceTableTester` into the .test.ts
